コミット仮
