+++ flying_kokaton.py
@@ -1,15 +1,12 @@
 import numpy as np
-import datetime
-import os
 import sys
-import uuid
 import pygame as pg
 import pygame
 from pygame.locals import *
-from module.name import Text, event_loop
+from module.name.name import Text, event_loop
+from module.scores.scores import Score, ScoreLogDAO
+from module.combos.combo import Combo
 from random import randint as ran
-from random import uniform
-from typing import List
 
 
 # 定数宣言部
@@ -18,7 +15,7 @@ WIDTH = 450
 RAD =25#こうかとんボールの半径
 BALL_X=75#ボールのⅹ距離
 BALL_Y=75
-TIMES  = 1000
+TIMES  = 480
 
 # 関数宣言部
 def elise(ball_lst: list,judge: list)-> list:
@@ -53,118 +50,6 @@ def drop_down(lis:list[list])->list[list]:
     return lis
 
 
-def jadge_combo(a, b):
-    if a == b or a + 10 == b or a == b + 10:
-        return True
-    else:
-        return False
-
-def jadge_double(lis:list[list], i:int, j:int, combo_type:int, combo_len:int = 3):
-    T = combo_len
-    if combo_type == 1:
-        for n in range(3):
-            for m in range(3):
-                if lis[i+n][j+m] >= 10:
-                    T -= 1
-
-    elif combo_type == 11:
-        for n in range(combo_len):
-            if lis[i][j+n] >= 10:
-                T -= 1
-
-    elif combo_type == 12:
-        for n in range(combo_len):
-            if lis[i+n][j] >= 10:
-                T -= 1
-
-    elif combo_type == 21:
-        if lis[i][j] >= 10 and lis[i][j+1] >= 10 and lis[i][j+2] >= 10 and lis[i-1][j+1] >= 10 and lis[i+1][j+1] >= 10:
-            T -= combo_len
-
-    elif combo_type == 22:
-        if lis[i][j] >= 10 and lis[i][j+2] >= 10 and lis[i+1][j] >= 10 and lis[i+1][j+1] >= 10 and lis[i+1][j+2] >= 10 and lis[i+2][j] >= 10 and lis[i+2][j+2] >= 10:
-            T -= combo_len
-    
-    elif combo_type == 23:
-        if lis[i][j] >= 10 and lis[i][j+2] >= 10 and lis[i+2][j] >= 10 and lis[i+2][j+2] >= 10 and lis[i][j+1] >= 10 and lis[i+2][j+1] >= 10 and lis[i+2][j+1] >= 10:
-            T -= combo_len
-
-    elif combo_type == 31:
-        for n in range(3):
-            if lis[i+n][j] >= 10:
-                    T -= 1
-                    if n == 2:
-                        for m in range(1, 3):
-                            if lis[i+n][j+m] >= 10:
-                                T -= 1
-
-    elif combo_type == 32:
-        for n in range(3):
-            if lis[i+n][j] >= 10:
-                    T -= 1
-                    if n == 2:
-                        for m in range(-2, 0):
-                            if lis[i+n][j+m] >= 10:
-                                T -= 1
-
-    elif combo_type == 33:
-        for n in range(3):
-            if lis[i+n][j] >= 10:
-                    T -= 1
-                    if n == 0:
-                        for m in range(1, 3):
-                            if lis[i+n][j+m] >= 10:
-                                T -= 1
-
-    elif combo_type == 34:
-        for n in range(3):
-            if lis[i][j+n] >= 10:
-                    T -= 1
-                    if n == 2:
-                        for m in range(1, 3):
-                            if lis[i+m][j+n] >= 10:
-                                T -= 1
-    elif combo_type == 41:
-        for n in range(3):
-            if lis[i][j+n] >= 10:
-                    T -= 1
-                    if n == 1:
-                        for m in range(1, 3):
-                            if lis[i+m][j+n] >= 10:
-                                T -= 1
-    elif combo_type == 42:
-        for n in range(3):
-            if lis[i+n][j] >= 10:
-                    T -= 1
-                    if n == 2:
-                        for m in range(-1, 2, 2):
-                            if lis[i+n][j+m] >= 10:
-                                T -= 1
-
-    elif combo_type == 43:
-        for n in range(3):
-            if lis[i+n][j] >= 10:
-                    T -= 1
-                    if n == 1:
-                        for m in range(1, 3):
-                            if lis[i+n][j+m] >= 10:
-                                T -= 1
-    
-    elif combo_type == 44:
-        for n in range(3):
-            if lis[i][j+n] >= 10:
-                    T -= 1
-                    if n == 2:
-                        for m in range(-1, 2, 2):
-                            if lis[i+m][j+n] >= 10:
-                                T -= 1
-
-    if T > 0:
-        return True
-    else:
-        return False
-
-
 
 # クラス宣言部
 class KoukatonDrop(pg.sprite.Sprite):
@@ -194,8 +79,6 @@ class KoukatonDrop(pg.sprite.Sprite):
         self.image.set_colorkey((0, 0, 0))
         self.rect = self.image.get_rect()
         self.rect.center = self.i*BALL_X+12,self.j*BALL_Y+215
-                
-
     
     def update(self,screen:pg.surface):
         screen.set_alpha(128)
@@ -206,141 +89,10 @@ class KoukatonDrop(pg.sprite.Sprite):
             screen.blit(self.kk_img, [self.rect.centerx, self.rect.centery])
         self.kill()
 
-class ScoreLogDAO:
-    """
-    それぞれのスコアデータの入出力を管理するクラス
-    TODO: できればデータベース化
-    """
-    def __init__(self,log_file_name:str = "score_log.csv", file_encoding:str = "utf-8") -> None:
-        """
-        スコアログを保存する先がなければ作成する
-        また、任意のファイル名へと変更できる
-        保存先のパスは"./logs"内になる
-
-        :param str log_file_name: ファイル名
-        :param str file_encoding: ファイルのエンコード方式
-        """
-        self.file_encoding = file_encoding
-        self.log_file_path = "./logs"
-
-        if not os.path.exists(self.log_file_path):
-            # フォルダがなければ作成する
-            os.mkdir(self.log_file_path)
-        
-        self.log_file = self.log_file_path + "/" + log_file_name
-        if not os.path.exists(self.log_file):
-            # ログファイルがなければ初期化する
-            with open(self.log_file, "w", encoding=self.file_encoding) as f:
-                f.write("uuid,player_name,score,created_time\n")
-
-        # タイムスタンプ準備
-        t_delta = datetime.timedelta(hours=9)
-        jts = datetime.timezone(t_delta, 'JST')
-        self.now = datetime.datetime.now(jts)
-    
-    def insert(self, uuid:str, player_name:str, score:int) -> bool:
-        """
-        プレイログを挿入する
-
-        :param str uuid: UUID
-        :param str player_name: プレイヤー名
-        :param str score:
-        :return: 成功したらTrueを返します
-        :rtype: bool
-        """
-        with open(self.log_file, "a", encoding=self.file_encoding) as f:
-            f.write(f"{uuid},{player_name},{score},{self.now.strftime('%Y/%m/%d %H:%M:%S')}\n")
-
-        return True
-    
-    def get(self)->list[tuple[str,str,int,str]]:
-        """
-        保存されているプレイログデータを取得します
-
-        :return: ログデータのtupleを返します
-        :rtype: tuple[tuple[str,str,int,str]]
-        """
-        result = []
-        with open(self.log_file, "r", encoding=self.file_encoding) as f:
-            f.readline()
-            result += [self.dismantling(row) for row in f]
-        
-        return result
-
-    def dismantling(self,row:str)->tuple[str,str,str,str]:
-        """
-        プレイログデータの一行をそれぞれの要素に分解し、tupleで返します
-
-        :return: uuid, player_name, score, created_time
-        :rtype: tuple[str,str,str,str]
-        """
-        datas = row.rstrip("\n").rsplit(",")
-        datas[2] = int(datas[2])
-
-        return tuple(datas)
-
-class Score:
-    """
-    スコア管理システム
-    """
-    def __init__(self, session:ScoreLogDAO, base_score:int = 1000, player_name:str = "guest"):
-        """
-        スコアをユーザと紐づけます
-        担当 : c0a23019
-        
-        :param str player_name: プレイヤー名
-        """
-        self.session = session
-
-        # スコア情報系
-        self.value = 0
-        self.player_name = player_name
-        self.player_uuid = str(uuid.uuid1())
-        self.base_score = base_score
-        # TODO: 遊んだ時間のlog取得
-
-        # 表示系
-        self.font = pg.font.Font(None, 50)
-        self.color = (0, 0, 255)
-        self.image = self.font.render(f"Score: {self.value}", 0, self.color)
-        self.rect = self.image.get_rect()
-        self.rect.center = 100, 100
-
-    def update(self, screen: pg.Surface):
-        """
-        スコア表示
-
-        :param Surface screen: スクリーン情報
-        """
-        self.image = self.font.render(f"Score: {self.value}", 0, self.color)
-        screen.blit(self.image, self.rect)
-
-    def add(self, add_score:int):
-        """
-        スコア加算
-
-        :param int add_score: 加算したい値
-        """
-        self.value += add_score
-
-    def calculate_combo_score(self, combo:int, bonus:float = 1.0):
-        """
-        コンボスコア計算
-        工夫点: ランダム性を持たせることで、ゲーム性を向上させる
-
-        :param int combo: 現在のコンボ数
-        :param float bonus: ボーナス倍率
-        """
-        combo_score = self.base_score * (combo ** 1.25) * (uniform(0.8, 1.2)) * bonus
-        self.add(round(combo_score))
-
-    def save(self) -> None:
-        """
-        スコアをファイルに保存する
-        """
-        self.session.insert(self.player_uuid, self.player_name, self.value)
-
 class Time_circulate():
+    """
+    タイマーに関わるクラス
+    """
     def __init__(self,past_time):
         self.past_time= past_time
         self.font=pg.font.SysFont("ha正楷書体pro",30)
@@ -355,9 +107,9 @@ class Time_circulate():
         if self.mode==0:
             txt = self.font.render(f"Operatinon time is 7second",True,(0,0,0))
         else:
-            txt = self.font.render(f"Operatinon limit {7-(tmr-self.past_time)//150}second",True,(0,0,0))
+            txt = self.font.render(f"Operatinon limit {7-(tmr-self.past_time)//60}second",True,(0,0,0))
         screen.blit(txt,[0,100])
-# クラス宣言部
+
 class PuzzleList():
     """
     パズル画面を管理するリストに関係するクラス
@@ -398,7 +150,7 @@ class PuzzleList():
         elif key == pg.K_LEFT and x > 0: # 左矢印キーが押されたときかつxがフレーム内
             x -= 1 # xを-1する
             if x < 0:
-                y = 0
+                x = 0
         elif key == pg.K_RIGHT and x < 6: # 右矢印キーが押されたときかつxがフレーム内
             x += 1 # xを+1する
             if x >5:
@@ -419,367 +171,28 @@ class PuzzleList():
                         break
         return array                        
 
-class Combo:
+class ComboLog:
     """
-    コンボの判定をするクラス
-    コンボの判定
-        1:横
-        2:縦
-    担当:瀬尾
+    combo表示に関わるクラス
     """
+    def __init__(self) -> None:
+        self.combo = 0
+        # 表示系
+        self.font = pg.font.Font(None, 50)
+        self.color = (0, 0, 255)
+        self.image = self.font.render(f"Now Combo: {self.combo}", 0, self.color)
+        self.rect = self.image.get_rect()
+        self.rect.center = 127, 55
     
-    combo_all = 0
-
-    def __init__(self, lis:list[list]):
-        self.combo_count = 0
-        self.lis = lis
-        check = 0
-        while True:
-            self.box_combo()
-            self.h_combo()
-            self.i_combo()
-            self.t_combo()
-            self.t_combo_rev()
-            self.t_left_combo()
-            self.t_right_combo()
-            self.l_combo_dl()
-            self.l_combo_dr()
-            self.l_combo_ul()
-            self.l_combo_ur()
-            self.cross_combo()
-            self.row_combo()
-            self.column_combo()
-            if check == self.combo_count:
-                break
-            check = self.combo_count
-        print("rep")
-        print(self.lis[0], "\n", self.lis[1], "\n", self.lis[2], "\n", self.lis[3], "\n", self.lis[4], "\n", self.lis[5])
-        self.elise(self.lis)
+    def add_combo(self, combo:int):
+        self.combo += combo
     
-    def box_combo(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 1, 9) and jadge_combo(self.lis[i][j], self.lis[i][j+1]) and jadge_combo(self.lis[i][j], self.lis[i][j+2]) and jadge_combo(self.lis[i][j], self.lis[i+1][j]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+2]) and jadge_combo(self.lis[i][j], self.lis[i+2][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+2]):
-                    stack += 2
-                    print("box", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 1, self.lis[i][j])
-                    break
+    def reset_combo(self):
+        self.combo = 0
 
-    def h_combo(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 22) and jadge_combo(self.lis[i][j], self.lis[i][j+2]) and jadge_combo(self.lis[i][j], self.lis[i+1][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+2]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+2]):
-                    stack += 2
-                    print("H", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 22, self.lis[i][j])
-                    break
-        
-    def i_combo(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 23) and jadge_combo(self.lis[i][j], self.lis[i][j+1]) and jadge_combo(self.lis[i][j], self.lis[i][j+2]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+2][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+2]):
-                    stack += 2
-                    print("I", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 23, self.lis[i][j])
-                    break
-
-    def t_combo(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 41, 5) and jadge_combo(self.lis[i][j], self.lis[i][j+1]) and jadge_combo(self.lis[i][j], self.lis[i][j+2]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+1]):
-                    stack += 2
-                    print("T", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 41, self.lis[i][j])
-                    break
-    
-    def t_combo_rev(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 42, 5) and jadge_combo(self.lis[i][j], self.lis[i+2][j-1]) and jadge_combo(self.lis[i][j], self.lis[i+2][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+1][j]):
-                    stack += 2
-                    print("T-rev", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 42, self.lis[i][j])
-                    break
-
-    def t_left_combo(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 43, 5) and jadge_combo(self.lis[i][j], self.lis[i+1][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+2]):
-                    stack += 2
-                    print("T-left", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 43, self.lis[i][j])
-                    break
-    
-    def t_right_combo(self):
-        for i in range(1, len(self.lis) - 1):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 44, 5) and jadge_combo(self.lis[i][j], self.lis[i][j+1]) and jadge_combo(self.lis[i][j], self.lis[i][j+2]) and jadge_combo(self.lis[i][j], self.lis[i-1][j+2]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+2]):
-                    stack += 2
-                    print("T-right", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 44, self.lis[i][j])
-                    break
-    
-    def l_combo_dl(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 31, 5) and jadge_combo(self.lis[i][j], self.lis[i+1][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+2]):
-                    stack += 2
-                    print("L-dl", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 31, self.lis[i][j])
-                    break
-
-    def l_combo_dr(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(2, len(self.lis)):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 32, 5) and jadge_combo(self.lis[i][j], self.lis[i+1][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j-1]) and jadge_combo(self.lis[i][j], self.lis[i+2][j-2]):
-                    stack += 2
-                    print("L-dr", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 32, self.lis[i][j])
-                    break
-    def l_combo_ul(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 33, 5) and jadge_combo(self.lis[i][j], self.lis[i+1][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j]) and jadge_combo(self.lis[i][j], self.lis[i][j+1]) and jadge_combo(self.lis[i][j], self.lis[i][j+2]):
-                    stack += 2
-                    print("L-ul", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 31, self.lis[i][j])
-                    break
-    def l_combo_ur(self):
-        for i in range(len(self.lis) - 2):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    continue
-                if jadge_double(self.lis, i, j, 34, 5) and jadge_combo(self.lis[i][j], self.lis[i][j+1]) and jadge_combo(self.lis[i][j], self.lis[i][j+2]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+2]) and jadge_combo(self.lis[i][j], self.lis[i+2][j+2]):
-                    stack += 2
-                    print("L-ur", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 34, self.lis[i][j])
-                    break
-    
-    def cross_combo(self):
-        for i in range(1, len(self.lis) - 1):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    stack -= 1
-                    continue
-                if jadge_double(self.lis, i, j, 21) and jadge_combo(self.lis[i][j], self.lis[i][j+1]) and jadge_combo(self.lis[i][j], self.lis[i][j+2]) and jadge_combo(self.lis[i][j], self.lis[i-1][j+1]) and jadge_combo(self.lis[i][j], self.lis[i+1][j+1]):
-                    stack += 2
-                    print("cross", [i, j])
-                    self.combo_add()
-                    self.change(self.lis, i, j, 21, self.lis[i][j])
-                    break
-
-    def row_combo(self):
-       for i in range(len(self.lis)):
-            stack = 0
-            for j in range(len(self.lis) - 2):
-                if stack > 0:
-                    stack -= 1
-                    continue
-                combo_len = 3
-                if jadge_combo(self.lis[i][j], self.lis[i][j+1]) and jadge_combo(self.lis[i][j], self.lis[i][j+2]):
-                    if j <= 2 and jadge_combo(self.lis[i][j], self.lis[i][j+3]):
-                        combo_len += 1
-                        if j <= 1 and jadge_combo(self.lis[i][j], self.lis[i][j+4]):
-                            combo_len += 1
-                            if j <= 0 and jadge_combo(self.lis[i][j], self.lis[i][j+5]):
-                                combo_len += 1
-                    if jadge_double(self.lis, i, j, 11, combo_len):
-                        self.combo_count += 1
-                        stack += combo_len-1
-                        print("row", [i, j], combo_len)
-                        self.combo_add()
-                        self.change(self.lis, i, j, 11, self.lis[i][j], combo_len)
-                        break
-                    
-    def column_combo(self):
-        for j in range(len(self.lis)):
-            stack = 0
-            for i in range(len(self.lis) - 2):
-                if stack > 0:
-                    stack -= 1
-                    continue
-                combo_len = 3
-                if jadge_combo(self.lis[i][j], self.lis[i+1][j]) and jadge_combo(self.lis[i][j], self.lis[i+2][j]):
-                    if i <= 2 and jadge_combo(self.lis[i][j], self.lis[i+3][j]):
-                        combo_len += 1
-                        if i <= 1 and self.lis[i][j] == self.lis[i+4][j]:
-                            combo_len += 1
-                            if i <= 0 and jadge_combo(self.lis[i][j], self.lis[i+5][j]):
-                                combo_len += 1
-                    if jadge_double(self.lis, i, j, 12, combo_len):
-                        self.combo_count += 1
-                        stack += combo_len-1
-                        print("column", [i, j], combo_len)
-                        self.combo_add()
-                        self.change(self.lis, i, j, 12, self.lis[i][j], combo_len)
-                        
-                        break
-
-    def get_count(self):
-        return self.combo_count
-    
-    def get_lis(self):
-        return self.lis
-    
-    def change(self, lis:list[list], i:int, j:int, combo_type:int, ele:int = 0, combo_len:int = 1):
-        #引数:lst　ボールの種類を保持するリスト
-        #コンボ判定されたball_lstを0にする
-        if ele >= 10:
-            ele = ele - 10
-        
-        if combo_type == 1:
-            for n in range(3):
-                for m in range(3):
-                    lis[i+n][j+m] = 10+ele
-
-        if combo_type == 11:
-            for n in range(combo_len):
-                lis[i][j+n] = 10+ele
-
-        elif combo_type == 12:
-            for n in range(combo_len):
-                lis[i+n][j] = 10+ele
-
-        elif combo_type == 21:
-            for n in range(3):
-                lis[i][j+n] = 10+ele
-            lis[i-1][j+1] = 10+ele
-            lis[i+1][j+1] = 10+ele
-
-        elif combo_type == 22:
-            for n in range(3):
-                lis[i+1][j+n] = 10+ele
-            lis[i][j] = 10+ele
-            lis[i][j+2] = 10+ele
-            lis[i+2][j] = 10+ele
-            lis[i+2][j+2] = 10+ele
-
-        elif combo_type == 23:
-                    for n in range(3):
-                        lis[i+n][j+1] = 10+ele
-                    lis[i][j] = 10+ele
-                    lis[i][j+2] = 10+ele
-                    lis[i+2][j] = 10+ele
-                    lis[i+2][j+2] = 10+ele
-
-        elif combo_type == 31:
-            for n in range(3):
-                lis[i+n][j] = 10+ele
-                if n == 2:
-                    for m in range(1, 3):
-                        lis[i+n][j+m] = 10+ele
-        
-        elif combo_type == 32:
-            for n in range(3):
-                lis[i+n][j] = 10+ele
-                if n == 2:
-                    for m in range(-2, 0):
-                        lis[i+n][j+m] = 10+ele
-        
-        elif combo_type == 33:
-            for n in range(3):
-                lis[i+n][j] = 10+ele
-                lis[i][j+n] = 10+ele
-
-        elif combo_type == 34:
-            for n in range(3):
-                lis[i][j+n] = 10+ele
-                if n == 2:
-                    for m in range(1, 3):
-                        lis[i+m][j+n] = 10+ele
-        
-        elif combo_type == 41:
-            for n in range(3):
-                lis[i][j+n] = 10+ele
-                if n == 1:
-                    for m in range(1, 3):
-                        lis[i+n][j] = 10+ele
-
-        elif combo_type == 42:
-            for n in range(3):
-                lis[i+n][j] = 10+ele
-                if n == 2:
-                    for m in range(-1, 2, 2):
-                        lis[i+n][j+m] = 10+ele
-
-        elif combo_type == 43:
-            for n in range(3):
-                lis[i+n][j] = 10+ele
-                if n == 1:
-                    for m in range(1, 3):
-                        lis[i+n][j+m] = 10+ele
-
-        elif combo_type == 44:
-            for n in range(3):
-                lis[i][j+n] = 10+ele
-                if n == 2:
-                    for m in range(-1, 2, 2):
-                        lis[i+m][j+n] = 10+ele
-
-        self.lis = lis
-
-    
-    def elise(self, lis:list[list]):
-        for i in range(len(lis)):
-            for j in range(len(lis[i])):
-                if lis[i][j] >= 10:
-                    lis[i][j] = 0
-        self.lis = lis
-    
-    @classmethod
-    def combo_add(self):
-        Combo.combo_all += 1
-        
-    @classmethod
-    def get_combo(self):
-        return Combo.combo_all
-    
-    @classmethod
-    def reset(self):
-        Combo.combo_all = 0
+    def update(self, screen: pg.Surface):
+        self.image = self.font.render(f"Now Combo: {self.combo}", 0, self.color)
+        screen.blit(self.image, self.rect)
 
 
 # main関数
@@ -803,6 +216,10 @@ def main():
     kk_rct.center = 300,200
     # ここまで
     score_log_DAO = ScoreLogDAO()
+    score = Score(score_log_DAO)
+    Combo.set_score(score)
+    Combo.set_screen(screen)
+    show_combo = ComboLog()
     drop_list_x = 0
     drop_list_y = 0
     change_list_X = 0
@@ -810,7 +227,9 @@ def main():
     text = Text()
     tmr = 0 # 時間保存
     tmrs=Time_circulate(tmr)
-
+    lis_m = PuzzleList()
+    lis = lis_m.get_lis()
+    
     ball = pg.sprite.Group()
 
     """
@@ -852,15 +271,12 @@ def main():
                 """
                 ゲームの初期化
                 """
-                lis= PuzzleList()
-                t = lis.get_lis()
-                for i in range(len(t)):
-                    for j in range(len(t[i])):
-                        ball.add(KoukatonDrop(lis.get_lis(),(i,j)))
+                for i in range(len(lis)):
+                    for j in range(len(lis[i])):
+                        ball.add(KoukatonDrop(lis,(i,j)))
                 status="game:1"
 
-            case "game:1":     
-                # 練習7
+            case "game:1":                    
                 for i in range(4):
                     screen.blit(bg_imgs[i%2], [-(tmr % 3200)+1600*i, 0])
                   
@@ -873,42 +289,66 @@ def main():
                         if event.key == pg.K_RETURN: # ENTERが押されたとき
                                 status = "game:2"
                                 tmrs.settime(tmr)
-                for i in range(len(t)):
-                    for j in range(len(t[i])):
-                        ball.add(KoukatonDrop(lis.get_lis(),(i,j)))
+                for i in range(len(lis)):
+                    for j in range(len(lis[i])):
+                        ball.add(KoukatonDrop(lis,(i,j)))
                     ball.update(screen)                               
                     ball.draw(screen)
                 screen.blit(ball_img,[drop_list_y*75+12,drop_list_x*75+215])
                 tmrs.set_mode(0)
-                tmrs.update(tmr,screen)       
+                tmrs.update(tmr,screen)    
+                score.update(screen)
+                show_combo.update(screen)
+
             case "game:2":
                 tmrs.set_mode(1)
+                show_combo.reset_combo() 
                 for event in event_list:
                     if event.type == pg.QUIT: return
                     elif event.type == pg.KEYDOWN:
                         change_list_X,change_list_Y=drop_list_x,drop_list_y
                         change_list_X,change_list_Y = PuzzleList.move_lect([change_list_X, change_list_Y], event.key)
                         if (change_list_X,change_list_Y) != (drop_list_x,drop_list_y): # X,Yとx,yの値が一致していないとき
-                            lis.get_lis()[change_list_X][change_list_Y],lis.get_lis()[drop_list_x][drop_list_y] = lis.get_lis()[drop_list_x][drop_list_y],lis.get_lis()[change_list_X][change_list_Y] # PuzzleListクラスのlisの中身を入れ替える
+                            lis[change_list_X][change_list_Y],lis[drop_list_x][drop_list_y] = lis[drop_list_x][drop_list_y],lis[change_list_X][change_list_Y] # PuzzleListクラスのlisの中身を入れ替える
                             drop_list_x,drop_list_y = change_list_X,change_list_Y
-                            print(lis.get_lis())
+                            print(lis)
                         if event.key == pg.K_RETURN: # ENTERが押されたとき
-                            status = "game:1"
-                print(tmr-tmrs.past_time)
+                            status = "game:3"
+                            
                 if tmr-tmrs.past_time>=TIMES:
-                    status="game:1"
+                    status = "game:3"
+
                 for i in range(4):
                     screen.blit(bg_imgs[i%2], [-(tmr % 3200)+1600*i, 0])
-                for i in range(len(t)):
-                    for j in range(len(t[i])):
-                        ball.add(KoukatonDrop(lis.get_lis(),(i,j)))
+                for i in range(len(lis)):
+                    for j in range(len(lis[i])):
+                        ball.add(KoukatonDrop(lis,(i,j)))
                     ball.update(screen)                               
                     ball.draw(screen)
                 screen.blit(ball_img,[drop_list_y*75+12,drop_list_x*75+215])
                 tmrs.update(tmr,screen)
+                score.update(screen)
+                show_combo.update(screen)
+
+            case "game:3":
+                check = Combo(lis)
+                co = check.get_count()
+                if co <= 0:
+                    status="game:1"
+                check = check.get_lis()
+                check = drop_down(check)
+                show_combo.add_combo(Combo.get_combo())
+                Combo.reset()
+                lis_m.set_lis(check)
+                print(check)
+                screen.blit(ball_img,[drop_list_y*75+12,drop_list_x*75+215])
+                tmrs.update(tmr,screen)
+                score.update(screen)
+                show_combo.update(screen)
+                
             
             case "log:0":
-                lis = score_log_DAO.get()
+                lis_log = score_log_DAO.get()
                 screen.fill((255, 255, 255))
                 rnk_img = pg.image.load("./ex5/fig/1278942.png")
                 rnk_rct = kk_img.get_rect()
@@ -920,24 +360,20 @@ def main():
 
                 screen.blit(rnk_img, rnk_rct)
                 font = pygame.font.Font(None, 20)
-                sor = sorted(lis, reverse=True, key=lambda x: x[3])
+                sor = sorted(lis_log, reverse=True, key=lambda x: x[3])
                 # スコア表示
                 for i, row in enumerate(sor):
                     screen.blit(font.render(str(row[1:]), True, (0,0,255)), (20, 50 + i*50))
                 status = "log:1"
             case "log:1":
                 for event in event_list:
-                    if event.type == pg.KEYDOWN:
-                        if event.key == pg.K_ESCAPE:
-                            status = "home:0"
-                
-
-
+                    if event.key == pg.K_ESCAPE:
+                        status = "home:0"
 
         # 共通処理部
         pg.display.update()
         tmr += 1        
-        clock.tick(200)
+        clock.tick(60)
 
 if __name__ == "__main__":
     pg.init()
