+++ module/name/__init__.py
@@ -1,0 +1,1 @@
+""" Name module. """
