+++ module/scores/__init__.py
@@ -1,0 +1,1 @@
+""" Score's Module """
